test: dedupe builder/validator apiStub helpers into @lodestar/test-utils
Move the byte-identical ApiClientStub type, httpClientStub, mockApiResponse,
and mockApiErrorResponse into a new @lodestar/test-utils/apiStub subpath so the
builder and validator suites share one source of truth. Each package keeps its
package-specific getApiClientStub() and re-exports the shared helpers, leaving
the consumer test files untouched.

Closes #9819.
