+++ packages/builder/test/unit/utils/apiStub.ts
@@ -1,9 +1,7 @@
-import {Mocked, vi} from "vitest";
-import {ApiClientMethods, ApiResponse, Endpoint, Endpoints, HttpStatusCode, IHttpClient} from "@lodestar/api";
+import {vi} from "vitest";
+import {ApiClientStub} from "@lodestar/test-utils/apiStub";
 
-export type ApiClientStub = {[K in keyof Endpoints]: Mocked<ApiClientMethods<Endpoints[K]>>} & {
-  httpClient: Mocked<IHttpClient>;
-};
+export {type ApiClientStub, mockApiErrorResponse, mockApiResponse} from "@lodestar/test-utils/apiStub";
 
 export function getApiClientStub(): ApiClientStub {
   return {
@@ -14,22 +12,4 @@ export function getApiClientStub(): ApiClientStub {
       getSyncingStatus: vi.fn(),
     },
   } as unknown as ApiClientStub;
-}
-
-export function mockApiResponse<T, M, E extends Endpoint<any, any, any, T, M>>({
-  data,
-  meta,
-}: (E["return"] extends void ? {data?: never} : {data: E["return"]}) &
-  (E["meta"] extends void ? {meta?: never} : {meta: E["meta"]})): ApiResponse<E> {
-  const response = new Response(null, {status: HttpStatusCode.OK});
-  const apiResponse = new ApiResponse<E>({} as any, null, response);
-  apiResponse.value = () => data as T;
-  apiResponse.meta = () => meta as M;
-  return apiResponse;
-}
-
-export async function mockApiErrorResponse<E extends Endpoint>(status: HttpStatusCode): Promise<ApiResponse<E>> {
-  const res = new ApiResponse<E>({} as any, null, new Response(null, {status}));
-  await res.errorBody();
-  return res;
 }
