+++ packages/test-utils/package.json
@@ -16,6 +16,11 @@
       "types": "./lib/index.d.ts",
       "import": "./lib/index.js"
     },
+    "./apiStub": {
+      "typescript": "./src/apiStub.ts",
+      "types": "./lib/apiStub.d.ts",
+      "import": "./lib/apiStub.js"
+    },
     "./doubles": {
       "typescript": "./src/doubles.ts",
       "types": "./lib/doubles.d.ts",
@@ -52,6 +57,7 @@
   "dependencies": {
     "@chainsafe/bls-keystore": "^3.1.0",
     "@chainsafe/blst": "^2.2.0",
+    "@lodestar/api": "workspace:^",
     "@lodestar/params": "workspace:^",
     "@lodestar/utils": "workspace:^",
     "axios": "^1.13.2",
