+++ packages/validator/test/utils/apiStub.ts
@@ -1,16 +1,7 @@
-import {Mocked, vi} from "vitest";
-import {ApiClientMethods, ApiResponse, Endpoint, Endpoints, HttpStatusCode, IHttpClient} from "@lodestar/api";
+import {vi} from "vitest";
+import {ApiClientStub, httpClientStub} from "@lodestar/test-utils/apiStub";
 
-type ApiClientStub = {[K in keyof Endpoints]: Mocked<ApiClientMethods<Endpoints[K]>>} & {
-  httpClient: Mocked<IHttpClient>;
-};
-
-const httpClientStub: IHttpClient = {
-  baseUrl: "",
-  request: vi.fn(),
-  urlsInits: [],
-  urlsScore: [],
-};
+export {mockApiErrorResponse, mockApiResponse} from "@lodestar/test-utils/apiStub";
 
 export function getApiClientStub(): ApiClientStub {
   return {
@@ -51,22 +42,4 @@ export function getApiClientStub(): ApiClientStub {
     },
     httpClient: httpClientStub,
   } as unknown as ApiClientStub;
-}
-
-export function mockApiResponse<T, M, E extends Endpoint<any, any, any, T, M>>({
-  data,
-  meta,
-}: (E["return"] extends void ? {data?: never} : {data: E["return"]}) &
-  (E["meta"] extends void ? {meta?: never} : {meta: E["meta"]})): ApiResponse<E> {
-  const response = new Response(null, {status: HttpStatusCode.OK});
-  const apiResponse = new ApiResponse<E>({} as any, null, response);
-  apiResponse.value = () => data as T;
-  apiResponse.meta = () => meta as M;
-  return apiResponse;
-}
-
-export async function mockApiErrorResponse<E extends Endpoint>(status: HttpStatusCode): Promise<ApiResponse<E>> {
-  const res = new ApiResponse<E>({} as any, null, new Response(null, {status}));
-  await res.errorBody();
-  return res;
 }
